% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_DATA_FLAGS}
\alias{GLOSSARY_OF_DATA_FLAGS}
\title{The common core reports data modification in several categories}
\format{
an object of class \code{tbl_df/tbl/data.frame} with
9 rows and 4 columns
\describe{
\item{DMS_FLAG}{\verb{<chr>} A short description of if and how the data were modified}
\item{Missing}{\verb{<lgl>} \code{TRUE} if numeric values are not available for these data}
\item{Tweaked}{\verb{<lgl>} \code{TRUE} if the reported value is different from the initial raw submitted value}
\item{Measured}{\verb{<lgl>} \code{TRUE} if the reported value matches a submitted value}
}
}
\source{
\url{https://nces.ed.gov/ccd/files.asp}
}
\usage{
GLOSSARY_OF_DATA_FLAGS
}
\description{
The common core reports data modification in several categories
}
\keyword{datasets}
