% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_GRADES}
\alias{GLOSSARY_OF_GRADES}
\title{The Common Core of Data uses long names for student grade levels}
\format{
an object of class \code{tbl_df/tbl/data.frame}, with
19 rows and 2 columns
\describe{
\item{Order}{\verb{<int>} the sort order of the grade label. Kindergarten is 0}
\item{GRADE}{\verb{<chr>} a long-ish name for the grade, e.g. "Grade 9"}
}
}
\source{
\url{https://nces.ed.gov/ccd/files.asp}
}
\usage{
GLOSSARY_OF_GRADES
}
\description{
The Common Core of Data uses long names for student grade levels
}
\keyword{datasets}
