% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_OMB_RACES}
\alias{GLOSSARY_OF_OMB_RACES}
\title{The Common Core of Data uses long names for racial/ethnic identities}
\format{
an object of class \code{tbl_df/tbl/data.frame}, with
9 rows and 2 columns
\describe{
\item{Census Letter}{\verb{<chr>} Some Census tables use a one-letter suffix, such as "D", to indicate that they pertain to a single OMB race}
\item{RACE_ETHNICITY}{\verb{<chr>} An OMB racial/ethnic identity, such as "Asian American".}
}
}
\source{
\url{https://nces.ed.gov/ccd/files.asp}
}
\usage{
GLOSSARY_OF_OMB_RACES
}
\description{
The Common Core of Data uses long names for racial/ethnic identities
}
\seealso{
\link[hercacstables:RACE_ETHNICITY_SUBTABLES]{hercacstables::RACE_ETHNICITY_SUBTABLES}
}
\keyword{datasets}
