% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_SEXES}
\alias{GLOSSARY_OF_SEXES}
\title{The Common Core of Data uses only two sexes}
\format{
an object of class \code{tbl_df/tbl/data.frame}, with
4 rows and 3 columns
\describe{
\item{SEX}{\verb{<chr>} "Female,"  "Male," or two different versions of "Other."}
\item{Non-Female}{\verb{<lgl>} \code{TRUE} if SEX is "Female," \code{FALSE} otherwise.}
\item{Non-Male}{\verb{<lgl>} \code{TRUE} if SEX is "Male," \code{FALSE} otherwise.}
}
}
\source{
\url{https://nces.ed.gov/ccd/files.asp}
}
\usage{
GLOSSARY_OF_SEXES
}
\description{
The Common Core of Data uses only two sexes
}
\keyword{datasets}
