% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GLOSSARY_OF_TOTAL_INDICATORS}
\alias{GLOSSARY_OF_TOTAL_INDICATORS}
\title{The Common Core of Data tables include specific subtotals}
\format{
an object of class \code{tbl_df/tbl/data.frame}, with
5 rows and 5 columns
\describe{
\item{By Race}{\verb{<lgl>} \code{TRUE} when each subtotal pertains to one OMB race.}
\item{By Sex}{\verb{<lgl>} \code{TRUE} when each subtotal pertains to one sex.}
\item{By Grade}{\verb{<lgl>} \code{TRUE} when each subtotal pertains to one grade.}
\item{Includes Adults}{\verb{<lgl>} \code{TRUE} when each subtotal includes adult students.}
\item{TOTAL_INDICATOR}{\verb{<chr>} a description of the subtotals' characteristics}
}
}
\source{
\url{https://nces.ed.gov/ccd/files.asp}
}
\usage{
GLOSSARY_OF_TOTAL_INDICATORS
}
\description{
The Common Core of Data tables include specific subtotals
}
\keyword{datasets}
